Rename extension quarkus-hibernate-search-orm-coordination-outbox-polling to quarkus-hibernate-search-orm-outbox-polling
To align on Hibernate Search renaming the module
hibernate-search-orm-coordination-outbox-polling to hibernate-search-orm-outbox-polling
