--- extensions/hibernate-search-orm-coordination-outbox-polling/deployment/src/test/java/io/quarkus/hibernate/search/orm/coordination/outboxpolling/test/configuration/defaultpu/package-info.java
@@ -1,4 +1,0 @@
-@PersistenceUnit(PersistenceUnit.DEFAULT)
-package io.quarkus.hibernate.search.orm.coordination.outboxpolling.test.configuration.defaultpu;
-
-import io.quarkus.hibernate.orm.PersistenceUnit;
