--- extensions/hibernate-search-orm-coordination-outbox-polling/deployment/src/test/java/io/quarkus/hibernate/search/orm/coordination/outboxpolling/test/configuration/pu1/package-info.java
@@ -1,4 +1,0 @@
-@PersistenceUnit("pu1")
-package io.quarkus.hibernate.search.orm.coordination.outboxpolling.test.configuration.pu1;
-
-import io.quarkus.hibernate.orm.PersistenceUnit;
