--- extensions/hibernate-search-orm-coordination-outbox-polling/runtime/src/main/java/io/quarkus/hibernate/search/orm/coordination/outboxpolling/runtime/HibernateSearchOutboxPollingRuntimeConfig.java
@@ -1,25 +1,0 @@
-package io.quarkus.hibernate.search.orm.coordination.outboxpolling.runtime;
-
-import java.util.Map;
-
-import io.quarkus.hibernate.orm.runtime.PersistenceUnitUtil;
-import io.quarkus.runtime.annotations.ConfigDocMapKey;
-import io.quarkus.runtime.annotations.ConfigPhase;
-import io.quarkus.runtime.annotations.ConfigRoot;
-import io.smallrye.config.ConfigMapping;
-import io.smallrye.config.WithParentName;
-import io.smallrye.config.WithUnnamedKey;
-
-@ConfigMapping(prefix = "quarkus.hibernate-search-orm")
-@ConfigRoot(phase = ConfigPhase.RUN_TIME)
-public interface HibernateSearchOutboxPollingRuntimeConfig {
-
-    /**
-     * Configuration for persistence units.
-     */
-    @WithParentName
-    @WithUnnamedKey(PersistenceUnitUtil.DEFAULT_PERSISTENCE_UNIT_NAME)
-    @ConfigDocMapKey("persistence-unit-name")
-    Map<String, HibernateSearchOutboxPollingRuntimeConfigPersistenceUnit> persistenceUnits();
-
-}
