+++ extensions/hibernate-search-orm-outbox-polling/deployment/src/test/java/io/quarkus/hibernate/search/orm/outboxpolling/test/configuration/defaultpu/package-info.java
@@ -1,0 +1,4 @@
+@PersistenceUnit(PersistenceUnit.DEFAULT)
+package io.quarkus.hibernate.search.orm.outboxpolling.test.configuration.defaultpu;
+
+import io.quarkus.hibernate.orm.PersistenceUnit;
