+++ extensions/hibernate-search-orm-outbox-polling/deployment/src/test/java/io/quarkus/hibernate/search/orm/outboxpolling/test/configuration/pu1/IndexedEntityForPU1.java
@@ -1,0 +1,27 @@
+package io.quarkus.hibernate.search.orm.outboxpolling.test.configuration.pu1;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+
+import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
+
+/**
+ * An indexed entity.
+ */
+@Entity
+@Indexed
+public class IndexedEntityForPU1 {
+
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+}
