+++ extensions/hibernate-search-orm-outbox-polling/deployment/src/test/java/io/quarkus/hibernate/search/orm/outboxpolling/test/configuration/pu1/package-info.java
@@ -1,0 +1,4 @@
+@PersistenceUnit("pu1")
+package io.quarkus.hibernate.search.orm.outboxpolling.test.configuration.pu1;
+
+import io.quarkus.hibernate.orm.PersistenceUnit;
