+++ extensions/hibernate-search-orm-outbox-polling/deployment/src/test/resources/application-multiple-persistence-units.properties
