+++ extensions/hibernate-search-orm-outbox-polling/deployment/src/test/resources/application.properties
