+++ extensions/pom.xml
@@ -96,7 +96,7 @@
         <module>hibernate-validator</module>
         <module>panache</module>
         <module>hibernate-search-orm-elasticsearch</module>
-        <module>hibernate-search-orm-coordination-outbox-polling</module>
+        <module>hibernate-search-orm-outbox-polling</module>
         <module>elasticsearch-rest-client-common</module>
         <module>elasticsearch-rest-client</module>
         <module>elasticsearch-java-client</module>
