+++ integration-tests/pom.xml
@@ -231,7 +231,7 @@
                 <module>hibernate-reactive-panache</module>
                 <module>hibernate-reactive-panache-kotlin</module>
                 <module>hibernate-search-orm-elasticsearch</module>
-                <module>hibernate-search-orm-elasticsearch-coordination-outbox-polling</module>
+                <module>hibernate-search-orm-elasticsearch-outbox-polling</module>
                 <module>hibernate-search-orm-opensearch</module>
                 <module>hibernate-search-orm-elasticsearch-tenancy</module>
                 <module>hibernate-orm-tenancy</module>
